perf(matrix): isolate storage metadata from client loading
